requirements.txt
torch==2.2.0
torchvision
torchaudio
rdkit
PyYAML
numpy<2.0
matplotlib
tqdm
scikit-learn
pandas